Replace select.py with select_bloq.py.
