+++ qualtran/bloqs/chemistry/sf/select_bloq_test.py
@@ -11,17 +11,8 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-from qualtran.bloqs.basic_gates import TGate
-from qualtran.bloqs.chemistry.sparse.select import _sel_sparse
+from qualtran.bloqs.chemistry.sf.select_bloq import _select
 
 
-def test_prep_inner(bloq_autotester):
-    bloq_autotester(_sel_sparse)
-
-
-def test_decompose_bloq_counts():
-    sel = _sel_sparse()
-    cost_decomp = sel.decompose_bloq().call_graph()[1][TGate()]
-    cost_call = sel.call_graph()[1][TGate()]
-    assert cost_call == cost_decomp
+def test_select(bloq_autotester):
+    bloq_autotester(_select)
